Move `RemoteImageCommentsLaoder` to production
